backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv
psycopg2-binary
django-mysql
PyMySQL
numpy
matplotlib
pandas
scikit-learn
torch==1.11.0
requests